requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiofauna==0.1.34
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiohttp-sse==2.1.0
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.76
botocore==1.29.76
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dataclasses-json==0.5.9
decorator==5.1.1
dnspython==2.3.0
frozenlist==1.3.3
future==0.18.3
geocoder==1.38.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.230
langchainplus-sdk==0.0.20
loguru==0.7.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.2
pydantic==1.10.11
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
ratelim==0.1.6
requests==2.31.0
rich==13.4.2
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
watchdog==3.0.0
wrapt==1.15.0
yarl==1.9.2
